+++ app/src/main/java/eu/faircode/email/FragmentOptionsSend.java
@@ -63,7 +63,6 @@ public class FragmentOptionsSend extends FragmentBase implements SharedPreferenc
     private SwitchCompat swSuggestReceived;
     private SwitchCompat swSuggestFrequently;
     private Button btnLocalContacts;
-    private Button btnBlockedSenders;
     private SwitchCompat swAutoIdentity;
     private SwitchCompat swSendChips;
     private SwitchCompat swSendReminders;
@@ -142,7 +141,6 @@ public class FragmentOptionsSend extends FragmentBase implements SharedPreferenc
         swSuggestReceived = view.findViewById(R.id.swSuggestReceived);
         swSuggestFrequently = view.findViewById(R.id.swSuggestFrequently);
         btnLocalContacts = view.findViewById(R.id.btnLocalContacts);
-        btnBlockedSenders = view.findViewById(R.id.btnBlockedSenders);
         swAutoIdentity = view.findViewById(R.id.swAutoIdentity);
         swSendChips = view.findViewById(R.id.swSendChips);
         swSendReminders = view.findViewById(R.id.swSendReminders);
@@ -263,15 +261,6 @@ public class FragmentOptionsSend extends FragmentBase implements SharedPreferenc
             public void onClick(View v) {
                 LocalBroadcastManager lbm = LocalBroadcastManager.getInstance(getContext());
                 lbm.sendBroadcast(new Intent(ActivitySetup.ACTION_MANAGE_LOCAL_CONTACTS));
-            }
-        });
-
-        btnBlockedSenders.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                LocalBroadcastManager lbm = LocalBroadcastManager.getInstance(getContext());
-                lbm.sendBroadcast(new Intent(ActivitySetup.ACTION_MANAGE_LOCAL_CONTACTS)
-                        .putExtra("junk", true));
             }
         });
