+++ app/src/main/res/layout/fragment_options_send.xml
@@ -195,19 +195,6 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/swSuggestFrequently" />
 
-                <Button
-                    android:id="@+id/btnBlockedSenders"
-                    style="?android:attr/buttonStyleSmall"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="12dp"
-                    android:layout_marginTop="12dp"
-                    android:drawableEnd="@drawable/twotone_report_24"
-                    android:drawablePadding="6dp"
-                    android:text="@string/title_blocked_senders"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/btnLocalContacts" />
-
                 <androidx.appcompat.widget.SwitchCompat
                     android:id="@+id/swAutoIdentity"
                     android:layout_width="0dp"
@@ -217,7 +204,7 @@
                     android:text="@string/title_advanced_auto_identity"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/btnBlockedSenders"
+                    app:layout_constraintTop_toBottomOf="@id/btnLocalContacts"
                     app:switchPadding="12dp" />
 
                 <eu.faircode.email.FixedTextView
