+++ app/src/main/res/layout/fragment_options_synchronize.xml
@@ -349,6 +349,17 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/cbDay5" />
 
+                <eu.faircode.email.FixedImageButton
+                    android:id="@+id/ibSchedules"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="12dp"
+                    android:contentDescription="@string/title_info"
+                    android:tooltipText="@string/title_info"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintTop_toBottomOf="@id/tvScheduleHint"
+                    app:srcCompat="@drawable/twotone_info_24" />
+
                 <eu.faircode.email.FixedTextView
                     android:id="@+id/tvScheduleIgnore"
                     android:layout_width="0dp"
@@ -364,16 +375,17 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/cbDay6" />
 
-                <eu.faircode.email.FixedImageButton
-                    android:id="@+id/ibSchedules"
+                <Button
+                    android:id="@+id/btnBlockedSenders"
+                    style="?android:attr/buttonStyleSmall"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="12dp"
-                    android:contentDescription="@string/title_info"
-                    android:tooltipText="@string/title_info"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/tvScheduleHint"
-                    app:srcCompat="@drawable/twotone_info_24" />
+                    android:drawableEnd="@drawable/twotone_report_24"
+                    android:drawablePadding="6dp"
+                    android:text="@string/title_blocked_senders"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toBottomOf="@id/tvScheduleIgnore" />
 
                 <androidx.constraintlayout.widget.Group
                     android:id="@+id/grpExempted"
